+++ .github/workflows/tests.yml
@@ -190,22 +190,6 @@ jobs:
           key: rspec-runtime-${{ matrix.build_type }}-${{ matrix.target }}-${{ github.run_id }}
           restore-keys: rspec-runtime-${{ matrix.build_type }}-${{ matrix.target }}-
 
-      - name: Check Zeitwerk eager_load
-        if: matrix.build_type == 'backend'
-        env:
-          LOAD_PLUGINS: ${{ (matrix.target == 'plugins') && '1' || '0' }}
-        run: |
-          if ! bin/rails zeitwerk:check --trace; then
-            echo
-            echo "---------------------------------------------"
-            echo
-            echo "::error::'bin/rails zeitwerk:check' failed - the app will fail to boot with 'eager_load=true' (e.g. in production)."
-            echo "To reproduce locally, run 'bin/rails zeitwerk:check'."
-            echo "Alternatively, you can run your local server/tests with the 'DISCOURSE_ZEITWERK_EAGER_LOAD=1' environment variable."
-            echo
-            exit 1
-          fi
-
       - name: Check Zeitwerk reloading
         if: matrix.build_type == 'backend'
         env:
