+++ app/controllers/session_controller.rb
@@ -132,6 +132,22 @@ class SessionController < ApplicationController
     end
   end
 
+  if Rails.env.test?
+    skip_before_action :check_xhr, only: :test_second_factor_restricted_route
+
+    def test_second_factor_restricted_route
+      result =
+        run_second_factor!(TestSecondFactorAction) do |manager|
+          manager.allow_backup_codes! if params[:allow_backup_codes]
+        end
+      if result.no_second_factors_enabled?
+        render json: { result: "no_second_factors_enabled" }
+      else
+        render json: { result: "second_factor_auth_completed" }
+      end
+    end
+  end
+
   def sso_login
     raise Discourse::NotFound unless SiteSetting.enable_discourse_connect
     raise Discourse::ReadOnly if @readonly_mode && !staff_writes_only_mode?
