+++ config/environments/test.rb
@@ -9,6 +9,12 @@ Discourse::Application.configure do
   # and recreated between test runs.  Don't rely on the data there!
   config.cache_classes = true
 
+  # Eager loading loads your entire application. When running a single test locally,
+  # this is usually not necessary, and can slow down your test suite. However, it's
+  # recommended that you enable it in continuous integration systems to ensure eager
+  # loading is working properly before deploying your code.
+  config.eager_load = ENV["CI"].present?
+
   # Configure static asset server for tests with Cache-Control for performance
   config.public_file_server.enabled = true
 
@@ -43,8 +49,6 @@ Discourse::Application.configure do
 
   config.assets.compile = true
   config.assets.digest = false
-
-  config.eager_load = ENV["DISCOURSE_ZEITWERK_EAGER_LOAD"] == "1"
 
   if ENV["RAILS_ENABLE_TEST_LOG"]
     config.logger = Logger.new(STDOUT)
